FAQ 등록/수정/삭제 권한 검증 강화

+++ src/main/java/zgoo/cpos/service/FaqService.java
@@ -14,12 +14,14 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import zgoo.cpos.domain.users.Faq;
 import zgoo.cpos.domain.users.Users;
+import zgoo.cpos.dto.menu.MenuAuthorityDto.MenuAuthorityBaseDto;
 import zgoo.cpos.dto.users.FaqDto;
 import zgoo.cpos.dto.users.FaqDto.FaqListDto;
 import zgoo.cpos.mapper.FaqMapper;
-import zgoo.cpos.mapper.NoticeMapper;
+import zgoo.cpos.repository.menu.MenuAuthorityRepository;
 import zgoo.cpos.repository.users.FaqRepository;
 import zgoo.cpos.repository.users.UsersRepository;
+import zgoo.cpos.util.MenuConstants;
 
 @Service
 @RequiredArgsConstructor
@@ -27,6 +29,7 @@ import zgoo.cpos.repository.users.UsersRepository;
 public class FaqService {
     private final FaqRepository faqRepository;
     private final UsersRepository usersRepository;
+    private final MenuAuthorityRepository menuAuthorityRepository;
 
     // FAQ 전체 조회
     public Page<FaqDto.FaqListDto> findFaqAll(int page, int size) {
@@ -138,11 +141,26 @@ public class FaqService {
 
     // FAQ 등록
     @Transactional
-    public void saveFaq(FaqDto.FaqRegDto dto) {
+    public void saveFaq(FaqDto.FaqRegDto dto, String loginUserId) {
         try {
-            Users users = this.usersRepository.finsUserOneNotJoinedComapny(dto.getUserId());
-            Faq faq = FaqMapper.toEntity(dto, users);
-            this.faqRepository.save(faq);
+            if (loginUserId == null || loginUserId.isEmpty()) {
+                throw new IllegalArgumentException("User ID is missing. Cannot save faq without login user ID."); 
+            }
+            dto.setUserId(loginUserId);
+
+            boolean isMod = checkSaveAuthority(loginUserId);
+
+            if (isMod) {
+                Users users = this.usersRepository.finsUserOneNotJoinedComapny(dto.getUserId());
+                Faq faq = FaqMapper.toEntity(dto, users);
+                this.faqRepository.save(faq);
+            } else {
+                log.warn("[saveFaq] Access without permission.");
+            }
+
+        } catch (IllegalArgumentException e) {
+            log.error("[saveFaq] Illegal argument error: {}", e.getMessage());
+            throw e;
         } catch (Exception e) {
             log.error("[saveFaq] error: {}", e.getMessage());
         }
@@ -150,17 +168,30 @@ public class FaqService {
 
     // FAQ 수정
     @Transactional
-    public FaqDto.FaqRegDto updateFaq(FaqDto.FaqRegDto dto) {
+    public FaqDto.FaqRegDto updateFaq(FaqDto.FaqRegDto dto, String loginUserId) {
         try {
-            Faq faq = this.faqRepository.findFaqOne(dto.getId());
+            if (loginUserId == null || loginUserId.isEmpty()) {
+                throw new IllegalArgumentException("User ID is missing. Cannot update faq without login user ID."); 
+            }
+            boolean isMod = checkUpdateAndDeleteAuthority(dto.getId(), loginUserId);
 
-            log.info("=== before update: {}", faq.toString());
+            if (isMod) {
+                Faq faq = this.faqRepository.findFaqOne(dto.getId());
 
-            faq.updateFaqInfo(dto);
+                log.info("=== before update: {}", faq.toString());
+    
+                faq.updateFaqInfo(dto);
+    
+                log.info("=== after update: {}", faq.toString());
+    
+                return FaqMapper.toDto(faq);
+            }
 
-            log.info("=== after update: {}", faq.toString());
-
-            return FaqMapper.toDto(faq);
+            log.warn("[updateFaq] Access without permission.");
+            return null;
+        } catch (IllegalArgumentException e) {
+            log.error("[updateFaq] Illegal argument error: {}", e.getMessage());
+            throw e;
         } catch (Exception e) {
             log.error("[updateFaq] error: {}", e.getMessage());
             return null;
@@ -169,8 +200,99 @@ public class FaqService {
 
     // FAQ 삭제
     @Transactional
-    public void deleteFaq(Long id) {
-        Long count = this.faqRepository.deleteFaqOne(id);
-        log.info("=== delete faq info: {}", count);
+    public void deleteFaq(Long id, String loginUserId) {
+        try {
+            if (loginUserId == null || loginUserId.isEmpty()) {
+                throw new IllegalArgumentException("User ID is missing. Cannot delete faq without login user ID."); 
+            }
+
+            boolean isMod = checkUpdateAndDeleteAuthority(id, loginUserId);
+
+            if (isMod) {
+                Long count = this.faqRepository.deleteFaqOne(id);
+                log.info("=== delete faq info: {}", count);
+            } else {
+                log.warn("[deleteFaq] Access without permission.");
+            }
+
+        } catch (IllegalArgumentException e) {
+            log.error("[deleteFaq] Illegal argument error: {}", e.getMessage());
+            throw e;
+        } catch (Exception e) {
+            log.error("[deleteFaq] error: {}", e.getMessage());
+        }
+
+    }
+
+    // edit & delete button control
+    public boolean buttonControl(Long id, String loginUserId) {
+        try {
+            Faq faq = this.faqRepository.findFaqOne(id);
+            String writer = faq.getUser().getUserId();
+
+            Users user = this.usersRepository.findUserOne(writer);
+            String userAuthority = user.getAuthority();
+
+            Users loginUser = this.usersRepository.findUserOne(loginUserId);
+            String loginUserAuthority = loginUser.getAuthority();
+
+            if (loginUserAuthority.equals("SU")) {
+                return true;
+            }
+
+            if (loginUserAuthority.equals("AD")) {
+                return !userAuthority.equals("SU");
+            }
+
+            return writer.equals(loginUserId);
+        } catch (Exception e) {
+            log.error("[FaqService >> buttonControl] error: {}", e.getMessage());
+            return false;
+        }
+    }
+
+    // faq save check
+    public boolean checkSaveAuthority(String loginUserId) {
+        Users loginUser = this.usersRepository.findUserOne(loginUserId);
+        String loginUserAuthority = loginUser.getAuthority();
+
+        if (loginUserAuthority.equals("SU")) {
+            log.info("[FaqService >> checkSaveAuthority] Supter Admin");
+            return true;
+        }
+
+        MenuAuthorityBaseDto dto = this.menuAuthorityRepository.findUserMenuAuthority(loginUser.getCompany().getId(),
+            loginUserAuthority, MenuConstants.FAQ);
+        return dto.getModYn().equals("Y");
+    }
+
+    // faq update & delete check
+    public boolean checkUpdateAndDeleteAuthority(Long id, String loginUserId) {
+        Users loginUser = this.usersRepository.findUserOne(loginUserId);
+        String loginUserAuthority = loginUser.getAuthority();
+
+        if (loginUserAuthority.equals("SU")) {
+            log.info("[FaqService >> checkUpdateAndDeleteAuthority] Super Admin");
+            return true;
+        }
+
+        Faq faq = this.faqRepository.findFaqOne(id);
+        String writer = faq.getUser().getUserId();
+        Users user = this.usersRepository.findUserOne(writer);
+        String userAuthority = user.getAuthority();
+
+        MenuAuthorityBaseDto dto = this.menuAuthorityRepository.findUserMenuAuthority(loginUser.getCompany().getId(),
+            loginUserAuthority, MenuConstants.FAQ);
+        String modYn = dto.getModYn();
+
+        if (modYn.equals("Y")) {
+            if (loginUserAuthority.equals("AD")) {
+                log.info("[FaqService >> checkUpdateAndDeleteAuthority] Admin");
+                return !userAuthority.equals("SU");
+            }
+            return writer.equals(loginUserId);
+        }
+        log.info("[FaqService >> checkUpdateAndDeleteAuthority] update & delete no permission");
+        return false;
     }
 }
